Deduplicate account validation and parsing
