+++ chains/solana/contracts/crates/common/Cargo.toml
@@ -1,12 +1,11 @@
 [package]
-name = "ccip_offramp"
+name = "common"
 version = "0.1.0-dev"
-description = "Created with Anchor"
 edition = "2021"
 
 [lib]
 crate-type = ["cdylib", "lib"]
-name = "ccip_offramp"
+name = "common"
 
 [features]
 no-entrypoint = []
@@ -19,11 +18,3 @@ default = []
 solana-program = "1.17.25" # pin solana to 1.17
 anchor-lang = { version = "0.29.0", features = ["init-if-needed"] }
 anchor-spl = "0.29.0"
-bytemuck = "1.7"
-ethnum = "1.5"
-fee_quoter = {path = "../fee-quoter", features = ["cpi"]}
-rmn_remote = {path = "../rmn-remote", features = ["cpi"]}
-common = {path = "../../crates/common"}
-
-[dev-dependencies]
-hex = "0.4.3"
