+++ chains/solana/contracts/programs/burnmint-token-pool/Cargo.toml
@@ -22,3 +22,4 @@ solana-program = "1.17.25" # pin solana to 1.17
 spl-math = { version = "0.2.0", features = [ "no-entrypoint" ] }
 base-token-pool = { version = "0.1.0-dev", path = "../base-token-pool/", features = ["no-entrypoint"] }
 rmn_remote = {path = "../rmn-remote", features = ["cpi"]}
+common = {path = "../../crates/common"}
