+++ chains/solana/contracts/programs/ccip-offramp/src/context.rs
@@ -1,5 +1,6 @@
 use anchor_lang::prelude::*;
 use bytemuck::{Pod, Zeroable};
+use common::seed;
 use solana_program::sysvar::instructions;
 
 use crate::messages::ExecutionReportSingleChain;
@@ -25,28 +26,6 @@ pub const fn valid_version(v: u8, max_version: u8) -> bool {
 
 pub const fn uninitialized(v: u8) -> bool {
     v == 0
-}
-
-/// Fixed seeds for PDA derivation: different context must use different seeds.
-pub mod seed {
-    pub const SOURCE_CHAIN: &[u8] = b"source_chain_state";
-    pub const COMMIT_REPORT: &[u8] = b"commit_report";
-    pub const CONFIG: &[u8] = b"config";
-    pub const REFERENCE_ADDRESSES: &[u8] = b"reference_addresses";
-    pub const STATE: &[u8] = b"state";
-
-    // arbitrary messaging signer
-    pub const EXTERNAL_EXECUTION_CONFIG: &[u8] = b"external_execution_config";
-    // token pool interaction signer
-    pub const EXTERNAL_TOKEN_POOL: &[u8] = b"external_token_pools_signer";
-    // signer for billing fee token transfer
-    pub const FEE_BILLING_SIGNER: &[u8] = b"fee_billing_signer";
-
-    // token specific
-    pub const TOKEN_ADMIN_REGISTRY: &[u8] = b"token_admin_registry";
-    pub const CCIP_TOKENPOOL_CONFIG: &[u8] = b"ccip_tokenpool_config";
-    pub const CCIP_TOKENPOOL_SIGNER: &[u8] = b"ccip_tokenpool_signer";
-    pub const TOKEN_POOL_CONFIG: &[u8] = b"ccip_tokenpool_chainconfig";
 }
 
 #[derive(Accounts)]
@@ -382,7 +361,7 @@ pub struct CommitReportContext<'info> {
     /// CHECK: fee quoter allowed price updater account, used to invoke fee quoter with price updates
     /// so that it can authorize the call made by this offramp
     #[account(
-        seeds = [fee_quoter::context::seed::ALLOWED_PRICE_UPDATER, fee_billing_signer.key().as_ref()],
+        seeds = [seed::ALLOWED_PRICE_UPDATER, fee_billing_signer.key().as_ref()],
         bump,
         seeds::program = fee_quoter.key(),
     )]
@@ -390,7 +369,7 @@ pub struct CommitReportContext<'info> {
 
     /// CHECK: fee quoter config account, used to invoke fee quoter with price updates
     #[account(
-        seeds = [fee_quoter::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = reference_addresses.load()?.fee_quoter,
     )]
@@ -407,7 +386,7 @@ pub struct CommitReportContext<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CURSES],
+        seeds = [seed::CURSES],
         bump,
         seeds::program = reference_addresses.load()?.rmn_remote,
     )]
@@ -415,7 +394,7 @@ pub struct CommitReportContext<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = reference_addresses.load()?.rmn_remote,
     )]
@@ -466,7 +445,7 @@ pub struct PriceOnlyCommitReportContext<'info> {
     /// CHECK: fee quoter allowed price updater account, used to invoke fee quoter with price updates
     /// so that it can authorize the call made by this offramp
     #[account(
-        seeds = [fee_quoter::context::seed::ALLOWED_PRICE_UPDATER, fee_billing_signer.key().as_ref()],
+        seeds = [seed::ALLOWED_PRICE_UPDATER, fee_billing_signer.key().as_ref()],
         bump,
         seeds::program = fee_quoter.key(),
     )]
@@ -474,7 +453,7 @@ pub struct PriceOnlyCommitReportContext<'info> {
 
     /// CHECK: fee quoter config account, used to invoke fee quoter with price updates
     #[account(
-        seeds = [fee_quoter::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = reference_addresses.load()?.fee_quoter,
     )]
@@ -491,7 +470,7 @@ pub struct PriceOnlyCommitReportContext<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CURSES],
+        seeds = [seed::CURSES],
         bump,
         seeds::program = reference_addresses.load()?.rmn_remote,
     )]
@@ -499,7 +478,7 @@ pub struct PriceOnlyCommitReportContext<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = reference_addresses.load()?.rmn_remote,
     )]
@@ -584,7 +563,7 @@ pub struct ExecuteReportContext<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CURSES],
+        seeds = [seed::CURSES],
         bump,
         seeds::program = reference_addresses.load()?.rmn_remote,
     )]
@@ -592,7 +571,7 @@ pub struct ExecuteReportContext<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = reference_addresses.load()?.rmn_remote,
     )]
