+++ chains/solana/contracts/programs/ccip-offramp/src/instructions/v1/execute.rs
@@ -1,8 +1,10 @@
 use anchor_lang::prelude::*;
+use common::seed;
+use common::v1::{validate_and_parse_token_accounts, TokenAccounts};
 use solana_program::instruction::Instruction;
 use solana_program::program::invoke_signed;
 
-use crate::context::{seed, ExecuteReportContext, OcrPluginType};
+use crate::context::{ExecuteReportContext, OcrPluginType};
 use crate::event::{ExecutionStateChanged, SkippedAlreadyExecutedMessage};
 use crate::instructions::interfaces::Execute;
 use crate::messages::{
@@ -16,8 +18,7 @@ use super::messages::{is_writable, Any2SVMMessage, ReleaseOrMintInV1, ReleaseOrM
 use super::ocr3base::{ocr3_transmit, ReportContext, Signatures};
 use super::ocr3impl::Ocr3ReportForExecutionReportSingleChain;
 use super::pools::{
-    calculate_token_pool_account_indices, get_balance, interact_with_pool,
-    validate_and_parse_token_accounts, TokenAccounts, CCIP_POOL_V1_RET_BYTES,
+    calculate_token_pool_account_indices, get_balance, interact_with_pool, CCIP_POOL_V1_RET_BYTES,
 };
 use super::rmn::verify_uncursed_cpi;
 
@@ -476,7 +477,7 @@ fn hash(
         &header_source_chain_selector,
         &header_dest_chain_selector,
         &on_ramp_address_size.to_be_bytes(),
-        &on_ramp_address.bytes(),
+        on_ramp_address.bytes(),
         // message header
         &msg.header.message_id,
         &msg.token_receiver.to_bytes(),
