+++ chains/solana/contracts/programs/ccip-router/src/context.rs
@@ -2,6 +2,7 @@ use anchor_lang::prelude::*;
 use anchor_spl::associated_token::get_associated_token_address_with_program_id;
 use anchor_spl::token::spl_token::native_mint;
 use anchor_spl::token_interface::{Mint, TokenAccount, TokenInterface};
+use common::seed;
 
 use crate::program::CcipRouter;
 use crate::state::{Config, Nonce};
@@ -22,25 +23,6 @@ pub const fn valid_version(v: u8, max_version: u8) -> bool {
 
 pub const fn uninitialized(v: u8) -> bool {
     v == 0
-}
-
-/// Fixed seeds for PDA derivation: different context must use different seeds.
-pub mod seed {
-    pub const DEST_CHAIN_STATE: &[u8] = b"dest_chain_state";
-    pub const NONCE: &[u8] = b"nonce";
-    pub const CONFIG: &[u8] = b"config";
-    pub const ALLOWED_OFFRAMP: &[u8] = b"allowed_offramp";
-
-    // token pool interaction signer
-    pub const EXTERNAL_TOKEN_POOL: &[u8] = b"external_token_pools_signer";
-    // signer for billing fee token transfer
-    pub const FEE_BILLING_SIGNER: &[u8] = b"fee_billing_signer";
-
-    // token specific
-    pub const TOKEN_ADMIN_REGISTRY: &[u8] = b"token_admin_registry";
-    pub const CCIP_TOKENPOOL_CONFIG: &[u8] = b"ccip_tokenpool_config";
-    pub const CCIP_TOKENPOOL_SIGNER: &[u8] = b"ccip_tokenpool_signer";
-    pub const TOKEN_POOL_CONFIG: &[u8] = b"ccip_tokenpool_chainconfig";
 }
 
 #[derive(Accounts)]
@@ -378,7 +360,7 @@ pub struct CcipSend<'info> {
 
     /// CHECK: This account is just used in the CPI to the Fee Quoter program
     #[account(
-        seeds = [fee_quoter::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = config.fee_quoter,
     )]
@@ -386,7 +368,7 @@ pub struct CcipSend<'info> {
 
     /// CHECK: This account is just used in the CPI to the Fee Quoter program
     #[account(
-        seeds = [fee_quoter::context::seed::DEST_CHAIN, destination_chain_selector.to_le_bytes().as_ref()],
+        seeds = [seed::DEST_CHAIN, destination_chain_selector.to_le_bytes().as_ref()],
         bump,
         seeds::program = config.fee_quoter,
     )]
@@ -394,7 +376,7 @@ pub struct CcipSend<'info> {
 
     /// CHECK: This account is just used in the CPI to the Fee Quoter program
     #[account(
-        seeds = [fee_quoter::context::seed::FEE_BILLING_TOKEN_CONFIG,
+        seeds = [seed::FEE_BILLING_TOKEN_CONFIG,
             if message.fee_token == Pubkey::default() {
                 native_mint::ID.as_ref() // pre-2022 WSOL
             } else {
@@ -408,7 +390,7 @@ pub struct CcipSend<'info> {
 
     /// CHECK: This account is just used in the CPI to the Fee Quoter program
     #[account(
-        seeds = [fee_quoter::context::seed::FEE_BILLING_TOKEN_CONFIG,
+        seeds = [seed::FEE_BILLING_TOKEN_CONFIG,
             config.link_token_mint.key().as_ref(),
         ],
         bump,
@@ -427,7 +409,7 @@ pub struct CcipSend<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CURSES],
+        seeds = [seed::CURSES],
         bump,
         seeds::program = config.rmn_remote,
     )]
@@ -435,7 +417,7 @@ pub struct CcipSend<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = config.rmn_remote,
     )]
@@ -491,7 +473,7 @@ pub struct GetFee<'info> {
 
     /// CHECK: This account is just used in the CPI to the Fee Quoter program
     #[account(
-        seeds = [fee_quoter::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = config.fee_quoter,
     )]
@@ -499,7 +481,7 @@ pub struct GetFee<'info> {
 
     /// CHECK: This account is just used in the CPI to the Fee Quoter program
     #[account(
-        seeds = [fee_quoter::context::seed::DEST_CHAIN, destination_chain_selector.to_le_bytes().as_ref()],
+        seeds = [seed::DEST_CHAIN, destination_chain_selector.to_le_bytes().as_ref()],
         bump,
         seeds::program = config.fee_quoter,
     )]
@@ -507,7 +489,7 @@ pub struct GetFee<'info> {
 
     /// CHECK: This account is just used in the CPI to the Fee Quoter program
     #[account(
-        seeds = [fee_quoter::context::seed::FEE_BILLING_TOKEN_CONFIG,
+        seeds = [seed::FEE_BILLING_TOKEN_CONFIG,
             if message.fee_token == Pubkey::default() {
                 native_mint::ID.as_ref() // pre-2022 WSOL
             } else {
@@ -521,7 +503,7 @@ pub struct GetFee<'info> {
 
     /// CHECK: This account is just used in the CPI to the Fee Quoter program
     #[account(
-        seeds = [fee_quoter::context::seed::FEE_BILLING_TOKEN_CONFIG,
+        seeds = [seed::FEE_BILLING_TOKEN_CONFIG,
             config.link_token_mint.key().as_ref(),
         ],
         bump,
