+++ chains/solana/contracts/programs/ccip-router/src/instructions/v1/token_admin_registry.rs
@@ -1,18 +1,17 @@
 use crate::{
     events::token_admin_registry as events,
     instructions::interfaces::TokenAdminRegistry,
+    token_admin_registry_writable,
     token_context::{
         OverridePendingTokenAdminRegistryByCCIPAdmin, OverridePendingTokenAdminRegistryByOwner,
     },
 };
 use anchor_lang::prelude::*;
 use anchor_spl::associated_token::get_associated_token_address_with_program_id;
+use common::seed;
 use solana_program::{address_lookup_table::state::AddressLookupTable, log::sol_log};
 
-use super::pools::token_admin_registry_writable;
-
 use crate::{
-    seed,
     token_context::{RegisterTokenAdminRegistryByCCIPAdmin, RegisterTokenAdminRegistryByOwner},
     AcceptAdminRoleTokenAdminRegistry, CcipRouterError, ModifyTokenAdminRegistry,
     SetPoolTokenAdminRegistry,
@@ -182,10 +181,7 @@ impl TokenAdminRegistry for Impl {
                 &pool_program,
             );
             let (fee_billing_config, _) = Pubkey::find_program_address(
-                &[
-                    fee_quoter::context::seed::FEE_BILLING_TOKEN_CONFIG,
-                    token_mint.as_ref(),
-                ],
+                &[seed::FEE_BILLING_TOKEN_CONFIG, token_mint.as_ref()],
                 &ctx.accounts.config.fee_quoter,
             );
