+++ chains/solana/contracts/programs/ccip-router/src/token_context.rs
@@ -1,4 +1,5 @@
 use anchor_lang::prelude::*;
+use common::seed;
 
 use crate::context::*;
 use crate::state::*;
@@ -180,4 +181,103 @@ pub struct AcceptAdminRoleTokenAdminRegistry<'info> {
     pub mint: InterfaceAccount<'info, Mint>, // underlying token that the pool wraps
     #[account(mut, address = token_admin_registry.pending_administrator @ CcipRouterError::Unauthorized)]
     pub authority: Signer<'info>,
+}
+
+pub mod token_admin_registry_writable {
+    use super::TokenAdminRegistry;
+
+    // set writable inserts bits from left to right
+    // index 0 is left-most bit
+    pub fn set(tar: &mut TokenAdminRegistry, index: u8) {
+        match index < 128 {
+            true => {
+                tar.writable_indexes[0] |= 1 << (127 - index);
+            }
+            false => {
+                tar.writable_indexes[1] |= 1 << (255 - index);
+            }
+        }
+    }
+
+    pub fn reset(tar: &mut TokenAdminRegistry) {
+        tar.writable_indexes = [0, 0];
+    }
+
+    #[cfg(test)]
+    mod tests {
+        use super::*;
+        use solana_program::pubkey::Pubkey;
+
+        fn is(tar: &TokenAdminRegistry, index: u8) -> bool {
+            match index < 128 {
+                true => tar.writable_indexes[0] & 1 << (127 - index) != 0,
+                false => tar.writable_indexes[1] & 1 << (255 - index) != 0,
+            }
+        }
+
+        #[test]
+        fn set_writable() {
+            let state = &mut TokenAdminRegistry {
+                version: 0,
+                administrator: Pubkey::default(),
+                pending_administrator: Pubkey::default(),
+                lookup_table: Pubkey::default(),
+                writable_indexes: [0, 0],
+                mint: Pubkey::default(),
+            };
+
+            set(state, 0);
+            set(state, 128);
+            assert_eq!(state.writable_indexes[0], 2u128.pow(127));
+            assert_eq!(state.writable_indexes[1], 2u128.pow(127));
+
+            reset(state);
+            assert_eq!(state.writable_indexes[0], 0);
+            assert_eq!(state.writable_indexes[1], 0);
+
+            set(state, 0);
+            set(state, 2);
+            set(state, 127);
+            set(state, 128);
+            set(state, 2 + 128);
+            set(state, 255);
+            assert_eq!(
+                state.writable_indexes[0],
+                2u128.pow(127) + 2u128.pow(127 - 2) + 2u128.pow(0)
+            );
+            assert_eq!(
+                state.writable_indexes[1],
+                2u128.pow(127) + 2u128.pow(127 - 2) + 2u128.pow(0)
+            );
+        }
+
+        #[test]
+        fn check_writable() {
+            let state = &TokenAdminRegistry {
+                version: 0,
+                administrator: Pubkey::default(),
+                pending_administrator: Pubkey::default(),
+                lookup_table: Pubkey::default(),
+                writable_indexes: [
+                    2u128.pow(127 - 7) + 2u128.pow(127 - 2) + 2u128.pow(127 - 4),
+                    2u128.pow(127 - 8) + 2u128.pow(127 - 56) + 2u128.pow(127 - 100),
+                ],
+                mint: Pubkey::default(),
+            };
+
+            assert!(!is(state, 0));
+            assert!(!is(state, 128));
+            assert!(!is(state, 255));
+
+            assert_eq!(state.writable_indexes[0].count_ones(), 3);
+            assert_eq!(state.writable_indexes[1].count_ones(), 3);
+
+            assert!(is(state, 7));
+            assert!(is(state, 2));
+            assert!(is(state, 4));
+            assert!(is(state, 128 + 8));
+            assert!(is(state, 128 + 56));
+            assert!(is(state, 128 + 100));
+        }
+    }
 }
