+++ chains/solana/contracts/programs/fee-quoter/Cargo.toml
@@ -21,6 +21,7 @@ anchor-lang = { version = "0.29.0", features = ["init-if-needed"] }
 anchor-spl = "0.29.0"
 bytemuck = "1.7"
 ethnum = "1.5"
+common = { path = "../../crates/common"}
 
 [dev-dependencies]
 hex = "0.4.3"
