+++ chains/solana/contracts/programs/test-token-pool/src/context.rs
@@ -4,6 +4,7 @@ use anchor_spl::{
     token_interface::{Mint, TokenAccount},
 };
 use base_token_pool::common::*;
+use common::seed;
 
 use crate::{ChainConfig, State};
 
@@ -120,7 +121,7 @@ pub struct TokenOfframp<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CURSES],
+        seeds = [seed::CURSES],
         bump,
         seeds::program = state.config.rmn_remote,
     )]
@@ -128,7 +129,7 @@ pub struct TokenOfframp<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = state.config.rmn_remote,
     )]
@@ -184,7 +185,7 @@ pub struct TokenOnramp<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CURSES],
+        seeds = [seed::CURSES],
         bump,
         seeds::program = state.config.rmn_remote,
     )]
@@ -192,7 +193,7 @@ pub struct TokenOnramp<'info> {
 
     /// CHECK: This account is just used in the CPI to the RMN Remote program
     #[account(
-        seeds = [rmn_remote::context::seed::CONFIG],
+        seeds = [seed::CONFIG],
         bump,
         seeds::program = state.config.rmn_remote,
     )]
